Add vgm2pcm
vgm2pcm is a program that converts a VGM file to the L16 format, as
described in RFC 3551. It takes 2 arguments: the input VGM file and the
output file. The output file may be a named pipe (FIFO), which makes it
possible to use vgm2pcm as a transcoding agent in MediaTomb.

Makefile was modified to support building vgm2pcm from the same objects
vgmplay uses. There is no longer a rule to make directories: they are
created automatically if necessary when building objects. The reason is
that when running 'make' to build vgmplay and vgm2pcm at the same time,
obj/vgm2pcm.o was created after vgmplay, and if 'make' is run again,
make detects that vgmplay needs to be rebuilt because the obj directory
was modified.